genome_analysis/requirements.txt
streamlit
Biopython
collections
datetime
pandas
os
tempfile
